Simplify RTK implementation to loading-only store
