+++ app/services/storeService/index.ts
@@ -1,3 +1,3 @@
-export * from './_articleSlice';
 export * from './_hooks';
+export * from './_loadingSlice';
 export * from './_store';
